Break out `config ...` commands into their own package
Continuing with the clean-up of `pkg/cmd/pulumi`, this factors out the
`config ...` hierarchy, along with shared functions that are used by
other commands for loading configuration.
