+++ pkg/cmd/pulumi/config/config.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package config
 
 import (
 	"context"
@@ -24,7 +24,6 @@ import (
 	"regexp"
 	"sort"
 	"strings"
-	"time"
 
 	"github.com/nbutton23/zxcvbn-go"
 	"github.com/spf13/cobra"
@@ -39,21 +38,17 @@ import (
 	cmdStack "github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/stack"
 	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/ui"
 	"github.com/pulumi/pulumi/pkg/v3/resource/stack"
-	"github.com/pulumi/pulumi/pkg/v3/secrets"
 	"github.com/pulumi/pulumi/pkg/v3/secrets/cloud"
 	"github.com/pulumi/pulumi/pkg/v3/secrets/passphrase"
 	pkgWorkspace "github.com/pulumi/pulumi/pkg/v3/workspace"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/apitype"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/diag/colors"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/resource/config"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/tokens"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/util/cmdutil"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/util/contract"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/util/logging"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/workspace"
 )
 
-func newConfigCmd() *cobra.Command {
+func NewConfigCmd() *cobra.Command {
 	var stack string
 	var showSecrets bool
 	var jsonOut bool
@@ -261,61 +256,6 @@ func newConfigCopyCmd(stack *string) *cobra.Command {
 	return cpCommand
 }
 
-func copySingleConfigKey(
-	ctx context.Context,
-	ssml cmdStack.SecretsManagerLoader,
-	configKey string,
-	path bool,
-	currentStack backend.Stack,
-	currentProjectStack *workspace.ProjectStack,
-	destinationStack backend.Stack,
-	destinationProjectStack *workspace.ProjectStack,
-) error {
-	var decrypter config.Decrypter
-	key, err := parseConfigKey(configKey)
-	if err != nil {
-		return fmt.Errorf("invalid configuration key: %w", err)
-	}
-
-	v, ok, err := currentProjectStack.Config.Get(key, path)
-	if err != nil {
-		return err
-	} else if !ok {
-		return fmt.Errorf("configuration key '%s' not found for stack '%s'", prettyKey(key), currentStack.Ref())
-	}
-
-	if v.Secure() {
-		var err error
-		var state cmdStack.SecretsManagerState
-		if decrypter, state, err = ssml.GetDecrypter(ctx, currentStack, currentProjectStack); err != nil {
-			return fmt.Errorf("could not create a decrypter: %w", err)
-		}
-		contract.Assertf(
-			state == cmdStack.SecretsManagerUnchanged,
-			"We're reading a secure value so the encryption information must be present already",
-		)
-	} else {
-		decrypter = config.NewPanicCrypter()
-	}
-
-	encrypter, _, cerr := ssml.GetEncrypter(ctx, destinationStack, destinationProjectStack)
-	if cerr != nil {
-		return cerr
-	}
-
-	val, err := v.Copy(decrypter, encrypter)
-	if err != nil {
-		return err
-	}
-
-	err = destinationProjectStack.Config.Set(key, val, path)
-	if err != nil {
-		return err
-	}
-
-	return cmdStack.SaveProjectStack(destinationStack, destinationProjectStack)
-}
-
 func newConfigGetCmd(stack *string) *cobra.Command {
 	var jsonOut bool
 	var open bool
@@ -350,7 +290,7 @@ func newConfigGetCmd(stack *string) *cobra.Command {
 				return err
 			}
 
-			key, err := parseConfigKey(args[0])
+			key, err := ParseConfigKey(args[0])
 			if err != nil {
 				return fmt.Errorf("invalid configuration key: %w", err)
 			}
@@ -409,7 +349,7 @@ func newConfigRmCmd(stack *string) *cobra.Command {
 				return err
 			}
 
-			key, err := parseConfigKey(args[0])
+			key, err := ParseConfigKey(args[0])
 			if err != nil {
 				return fmt.Errorf("invalid configuration key: %w", err)
 			}
@@ -477,7 +417,7 @@ func newConfigRmAllCmd(stack *string) *cobra.Command {
 			}
 
 			for _, arg := range args {
-				key, err := parseConfigKey(arg)
+				key, err := ParseConfigKey(arg)
 				if err != nil {
 					return fmt.Errorf("invalid configuration key: %w", err)
 				}
@@ -656,7 +596,7 @@ func newConfigSetCmd(stack *string) *cobra.Command {
 				return err
 			}
 
-			key, err := parseConfigKey(args[0])
+			key, err := ParseConfigKey(args[0])
 			if err != nil {
 				return fmt.Errorf("invalid configuration key: %w", err)
 			}
@@ -845,61 +785,6 @@ func newConfigSetAllCmd(stack *string) *cobra.Command {
 	return setCmd
 }
 
-func parseKeyValuePair(pair string) (config.Key, string, error) {
-	// Split the arg on the first '=' to separate key and value.
-	splitArg := strings.SplitN(pair, "=", 2)
-
-	// Check if the key is wrapped in quote marks and split on the '=' following the wrapping quote.
-	firstChar := string([]rune(pair)[0])
-	if firstChar == "\"" || firstChar == "'" {
-		pair = strings.TrimPrefix(pair, firstChar)
-		splitArg = strings.SplitN(pair, firstChar+"=", 2)
-	}
-
-	if len(splitArg) < 2 {
-		return config.Key{}, "", errors.New("config value must be in the form [key]=[value]")
-	}
-	key, err := parseConfigKey(splitArg[0])
-	if err != nil {
-		return config.Key{}, "", fmt.Errorf("invalid configuration key: %w", err)
-	}
-
-	value := splitArg[1]
-	return key, value, nil
-}
-
-func parseConfigKey(key string) (config.Key, error) {
-	// As a convenience, we'll treat any key with no delimiter as if:
-	// <program-name>:<key> had been written instead
-	if !strings.Contains(key, tokens.TokenDelimiter) {
-		proj, err := workspace.DetectProject()
-		if err != nil {
-			return config.Key{}, err
-		}
-
-		return config.ParseKey(fmt.Sprintf("%s:%s", proj.Name, key))
-	}
-
-	return config.ParseKey(key)
-}
-
-func prettyKey(k config.Key) string {
-	proj, err := workspace.DetectProject()
-	if err != nil {
-		return fmt.Sprintf("%s:%s", k.Namespace(), k.Name())
-	}
-
-	return prettyKeyForProject(k, proj)
-}
-
-func prettyKeyForProject(k config.Key, proj *workspace.Project) string {
-	if k.Namespace() == string(proj.Name) {
-		return k.Name()
-	}
-
-	return fmt.Sprintf("%s:%s", k.Namespace(), k.Name())
-}
-
 // configValueJSON is the shape of the --json output for a configuration value.  While we can add fields to this
 // structure in the future, we should not change existing fields.
 type configValueJSON struct {
@@ -1032,7 +917,7 @@ func listConfig(
 				return fmt.Errorf("could not decrypt configuration value: %w", err)
 			}
 
-			rows = append(rows, cmdutil.TableRow{Columns: []string{prettyKey(key), decrypted}})
+			rows = append(rows, cmdutil.TableRow{Columns: []string{PrettyKey(key), decrypted}})
 		}
 
 		ui.FprintTable(stdout, cmdutil.Table{
@@ -1200,7 +1085,7 @@ func getConfig(
 		return nil
 	}
 
-	return fmt.Errorf("configuration key '%s' not found for stack '%s'", prettyKey(key), stack.Ref())
+	return fmt.Errorf("configuration key '%s' not found for stack '%s'", PrettyKey(key), stack.Ref())
 }
 
 // keyPattern is the regular expression a configuration key must match before we check (and error) if we think
@@ -1235,84 +1120,6 @@ func looksLikeSecret(k config.Key, v string) bool {
 		(info.Entropy >= (entropyThreshold/2) && entropyPerChar >= entropyPerCharThreshold)
 }
 
-func getAndSaveSecretsManager(
-	ctx context.Context,
-	ssml cmdStack.SecretsManagerLoader,
-	stack backend.Stack,
-	workspaceStack *workspace.ProjectStack,
-) (secrets.Manager, error) {
-	sm, state, err := ssml.GetSecretsManager(ctx, stack, workspaceStack)
-	if err != nil {
-		return nil, fmt.Errorf("get stack secrets manager: %w", err)
-	}
-	if state != cmdStack.SecretsManagerUnchanged {
-		if err = cmdStack.SaveProjectStack(stack, workspaceStack); err != nil && state == cmdStack.SecretsManagerMustSave {
-			return nil, fmt.Errorf("save stack config: %w", err)
-		}
-	}
-	return sm, nil
-}
-
-// Attempts to load configuration for the given stack.
-func getStackConfiguration(
-	ctx context.Context,
-	ssml cmdStack.SecretsManagerLoader,
-	stack backend.Stack,
-	project *workspace.Project,
-) (backend.StackConfiguration, secrets.Manager, error) {
-	return getStackConfigurationWithFallback(ctx, ssml, stack, project, nil)
-}
-
-// getStackConfigurationOrLatest attempts to load a current stack configuration
-// using getStackConfiguration. If that fails due to not being run within a
-// valid project, the latest configuration from the backend is returned. This is
-// primarily for use in commands like `pulumi destroy`, where it is useful to be
-// able to clean up a stack whose configuration has already been deleted as part
-// of that cleanup.
-func getStackConfigurationOrLatest(
-	ctx context.Context,
-	ssml cmdStack.SecretsManagerLoader,
-	stack backend.Stack,
-	project *workspace.Project,
-) (backend.StackConfiguration, secrets.Manager, error) {
-	return getStackConfigurationWithFallback(
-		ctx, ssml, stack, project,
-		func(err error) (config.Map, error) {
-			if errors.Is(err, workspace.ErrProjectNotFound) {
-				// This error indicates that we're not being run in a project directory.
-				// We should fallback on the backend.
-				return backend.GetLatestConfiguration(ctx, stack)
-			}
-			return nil, err
-		})
-}
-
-func needsCrypter(cfg config.Map, env esc.Value) bool {
-	var hasSecrets func(v esc.Value) bool
-	hasSecrets = func(v esc.Value) bool {
-		if v.Secret {
-			return true
-		}
-		switch v := v.Value.(type) {
-		case []esc.Value:
-			for _, v := range v {
-				if hasSecrets(v) {
-					return true
-				}
-			}
-		case map[string]esc.Value:
-			for _, v := range v {
-				if hasSecrets(v) {
-					return true
-				}
-			}
-		}
-		return false
-	}
-
-	return cfg.HasSecureValue() || hasSecrets(env)
-}
-
 func checkStackEnv(
 	ctx context.Context,
 	stack backend.Stack,
@@ -1336,128 +1143,6 @@ func checkStackEnv(
 	return envs.CheckYAMLEnvironment(ctx, orgName, yaml)
 }
 
-func openStackEnv(
-	ctx context.Context,
-	stack backend.Stack,
-	workspaceStack *workspace.ProjectStack,
-) (*esc.Environment, []apitype.EnvironmentDiagnostic, error) {
-	yaml := workspaceStack.EnvironmentBytes()
-	if len(yaml) == 0 {
-		return nil, nil, nil
-	}
-
-	envs, ok := stack.Backend().(backend.EnvironmentsBackend)
-	if !ok {
-		return nil, nil, fmt.Errorf("backend %v does not support environments", stack.Backend().Name())
-	}
-	orgNamer, ok := stack.(interface{ OrgName() string })
-	if !ok {
-		return nil, nil, fmt.Errorf("cannot determine organzation for stack %v", stack.Ref())
-	}
-	orgName := orgNamer.OrgName()
-
-	return envs.OpenYAMLEnvironment(ctx, orgName, yaml, 2*time.Hour)
-}
-
-func getStackConfigurationWithFallback(
-	ctx context.Context,
-	ssml cmdStack.SecretsManagerLoader,
-	s backend.Stack,
-	project *workspace.Project,
-	fallbackGetConfig func(err error) (config.Map, error), // optional
-) (backend.StackConfiguration, secrets.Manager, error) {
-	workspaceStack, err := cmdStack.LoadProjectStack(project, s)
-	if err != nil || workspaceStack == nil {
-		if fallbackGetConfig == nil {
-			return backend.StackConfiguration{}, nil, err
-		}
-		// On first run or the latest configuration is unavailable, fallback to check the project's configuration
-		cfg, err := fallbackGetConfig(err)
-		if err != nil {
-			return backend.StackConfiguration{}, nil, fmt.Errorf(
-				"stack configuration could not be loaded from either Pulumi.yaml or the backend: %w", err)
-		}
-		workspaceStack = &workspace.ProjectStack{
-			Config: cfg,
-		}
-	}
-
-	sm, err := getAndSaveSecretsManager(ctx, ssml, s, workspaceStack)
-	if err != nil {
-		return backend.StackConfiguration{}, nil, err
-	}
-
-	config, err := getStackConfigurationFromProjectStack(ctx, s, project, sm, workspaceStack)
-	if err != nil {
-		return backend.StackConfiguration{}, nil, err
-	}
-	return config, sm, nil
-}
-
-func getStackConfigurationFromProjectStack(
-	ctx context.Context,
-	stack backend.Stack,
-	project *workspace.Project,
-	sm secrets.Manager,
-	workspaceStack *workspace.ProjectStack,
-) (backend.StackConfiguration, error) {
-	env, diags, err := openStackEnv(ctx, stack, workspaceStack)
-	if err != nil {
-		return backend.StackConfiguration{}, fmt.Errorf("opening environment: %w", err)
-	}
-	if len(diags) != 0 {
-		printESCDiagnostics(os.Stderr, diags)
-		return backend.StackConfiguration{}, errors.New("opening environment: too many errors")
-	}
-
-	var pulumiEnv esc.Value
-	if env != nil {
-		warnOnNoEnvironmentEffects(os.Stdout, env)
-
-		pulumiEnv = env.Properties["pulumiConfig"]
-
-		_, environ, secrets, err := cli.PrepareEnvironment(env, nil)
-		if err != nil {
-			return backend.StackConfiguration{}, fmt.Errorf("preparing environment: %w", err)
-		}
-		if len(secrets) != 0 {
-			logging.AddGlobalFilter(logging.CreateFilter(secrets, "[secret]"))
-		}
-
-		for _, kvp := range environ {
-			if name, value, ok := strings.Cut(kvp, "="); ok {
-				if err := os.Setenv(name, value); err != nil {
-					return backend.StackConfiguration{}, fmt.Errorf("setting environment variable %v: %w", name, err)
-				}
-			}
-		}
-	}
-
-	// If there are no secrets in the configuration, we should never use the decrypter, so it is safe to return
-	// one which panics if it is used. This provides for some nice UX in the common case (since, for example, building
-	// the correct decrypter for the diy backend would involve prompting for a passphrase)
-	if !needsCrypter(workspaceStack.Config, pulumiEnv) {
-		return backend.StackConfiguration{
-			EnvironmentImports: workspaceStack.Environment.Imports(),
-			Environment:        pulumiEnv,
-			Config:             workspaceStack.Config,
-			Decrypter:          config.NewPanicCrypter(),
-		}, nil
-	}
-
-	crypter, err := sm.Decrypter()
-	if err != nil {
-		return backend.StackConfiguration{}, fmt.Errorf("getting configuration decrypter: %w", err)
-	}
-
-	return backend.StackConfiguration{
-		EnvironmentImports: workspaceStack.Environment.Imports(),
-		Environment:        pulumiEnv,
-		Config:             workspaceStack.Config,
-		Decrypter:          crypter,
-	}, nil
-}
-
 func warnOnNoEnvironmentEffects(out io.Writer, env *esc.Environment) {
 	hasEnvVars := len(env.GetEnvironmentVariables()) != 0
 	hasFiles := len(env.GetTemporaryFiles()) != 0
@@ -1469,5 +1154,18 @@ func warnOnNoEnvironmentEffects(out io.Writer, env *esc.Environment) {
 		fmt.Fprintln(out, color.Colorize(colors.SpecWarning+"The stack's environment does not define the `environmentVariables`, `files`, or `pulumiConfig` properties."))
 		fmt.Fprintln(out, color.Colorize(colors.SpecWarning+"Without at least one of these properties, the environment will not affect the stack's behavior."+colors.Reset))
 		fmt.Fprintln(out)
+	}
+}
+
+func printESCDiagnostics(out io.Writer, diags []apitype.EnvironmentDiagnostic) {
+	for _, d := range diags {
+		if d.Range != nil {
+			fmt.Fprintf(out, "%v:", d.Range.Environment)
+			if d.Range.Begin.Line != 0 {
+				fmt.Fprintf(out, "%v:%v:", d.Range.Begin.Line, d.Range.Begin.Column)
+			}
+			fmt.Fprintf(out, " ")
+		}
+		fmt.Fprintf(out, "%v\n", d.Summary)
 	}
 }
