+++ pkg/cmd/pulumi/config/config_env_ls.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2023, Pulumi Corporation.
+// Copyright 2016-2024, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,52 +12,45 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package config
 
 import (
 	"context"
 
 	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/cmd"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/util/cmdutil"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/workspace"
 	"github.com/spf13/cobra"
 )
 
-func newConfigEnvRmCmd(parent *configEnvCmd) *cobra.Command {
-	impl := configEnvRmCmd{parent: parent}
+func newConfigEnvLsCmd(parent *configEnvCmd) *cobra.Command {
+	var jsonOut bool
+
+	impl := configEnvLsCmd{parent: parent, jsonOut: &jsonOut}
 
 	cmd := &cobra.Command{
-		Use:   "rm <environment-name>",
-		Short: "Remove environment from a stack",
-		Long:  "Removes an environment from a stack's import list.",
-		Args:  cmdutil.ExactArgs(1),
+		Use:   "ls",
+		Short: "Lists imported environments.",
+		Long:  "Lists the environments imported into a stack's configuration.",
+		Args:  cmdutil.NoArgs,
 		Run: cmd.RunCmdFunc(func(cmd *cobra.Command, args []string) error {
 			parent.initArgs()
 			return impl.run(cmd.Context(), args)
 		}),
 	}
 
-	cmd.Flags().BoolVar(
-		&impl.showSecrets, "show-secrets", false,
-		"Show secret values in plaintext instead of ciphertext")
 	cmd.Flags().BoolVarP(
-		&impl.yes, "yes", "y", false,
-		"True to save changes without prompting")
+		&jsonOut, "json", "j", false,
+		"Emit output as JSON")
 
 	return cmd
 }
 
-type configEnvRmCmd struct {
+type configEnvLsCmd struct {
 	parent *configEnvCmd
 
-	showSecrets bool
-	yes         bool
+	jsonOut *bool
 }
 
-func (cmd *configEnvRmCmd) run(ctx context.Context, args []string) error {
-	return cmd.parent.editStackEnvironment(
-		ctx, cmd.showSecrets, cmd.yes, func(stack *workspace.ProjectStack) error {
-			stack.Environment = stack.Environment.Remove(args[0])
-			return nil
-		})
+func (cmd *configEnvLsCmd) run(ctx context.Context, _ []string) error {
+	return cmd.parent.listStackEnvironments(ctx, *cmd.jsonOut)
 }
