+++ pkg/cmd/pulumi/config/config_env_rm_test.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2023, Pulumi Corporation.
+// Copyright 2016-2024, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package config
 
 import (
 	"bytes"
@@ -25,7 +25,7 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestConfigEnvAddCmd(t *testing.T) {
+func TestConfigEnvRmCmd(t *testing.T) {
 	t.Parallel()
 
 	projectYAML := `name: test
@@ -34,34 +34,23 @@ runtime: yaml`
 	t.Run("no imports", func(t *testing.T) {
 		t.Parallel()
 
-		env := &esc.Environment{
-			Properties: map[string]esc.Value{
-				"pulumiConfig": esc.NewValue(map[string]esc.Value{
-					"aws:region": esc.NewValue("us-west-2"),
-				}),
-			},
-		}
-
 		var newStackYAML string
 		stdin := strings.NewReader("y")
 		var stdout bytes.Buffer
-		parent := newConfigEnvCmdForTest(stdin, &stdout, projectYAML, "", env, nil, &newStackYAML)
-		add := &configEnvAddCmd{parent: parent}
+		parent := newConfigEnvCmdForTest(stdin, &stdout, projectYAML, "", nil, nil, &newStackYAML)
+		rm := &configEnvRmCmd{parent: parent}
 		ctx := context.Background()
-		err := add.run(ctx, []string{"env"})
+		err := rm.run(ctx, []string{"env"})
 		require.NoError(t, err)
 
-		const expectedOut = `KEY         VALUE
-aws:region  us-west-2
+		const expectedOut = `KEY  VALUE
 
 Save? Yes
 `
 
 		assert.Equal(t, expectedOut, cleanStdoutIncludingPrompt(stdout.String()))
 
-		const expectedYAML = `environment:
-  - env
-`
+		const expectedYAML = "{}\n"
 
 		assert.Equal(t, expectedYAML, newStackYAML)
 	})
@@ -69,48 +58,65 @@ Save? Yes
 	t.Run("no imports, yes", func(t *testing.T) {
 		t.Parallel()
 
-		env := &esc.Environment{
-			Properties: map[string]esc.Value{
-				"pulumiConfig": esc.NewValue(map[string]esc.Value{
-					"aws:region": esc.NewValue("us-west-2"),
-				}),
-			},
-		}
-
 		var newStackYAML string
-		stdin := strings.NewReader("y")
 		var stdout bytes.Buffer
-		parent := newConfigEnvCmdForTest(stdin, &stdout, projectYAML, "", env, nil, &newStackYAML)
-		add := &configEnvAddCmd{parent: parent, yes: true}
+		parent := newConfigEnvCmdForTest(nil, &stdout, projectYAML, "", nil, nil, &newStackYAML)
+		rm := &configEnvRmCmd{parent: parent, yes: true}
 		ctx := context.Background()
-		err := add.run(ctx, []string{"env"})
+		err := rm.run(ctx, []string{"env"})
 		require.NoError(t, err)
 
-		const expectedOut = `KEY         VALUE
-aws:region  us-west-2
-`
+		const expectedOut = "KEY  VALUE\n"
 
 		assert.Equal(t, expectedOut, cleanStdoutIncludingPrompt(stdout.String()))
 
-		const expectedYAML = `environment:
-  - env
-`
+		const expectedYAML = "{}\n"
 
 		assert.Equal(t, expectedYAML, newStackYAML)
 	})
 
-	t.Run("no effects", func(t *testing.T) {
+	t.Run("one import", func(t *testing.T) {
 		t.Parallel()
+
+		const stackYAML = `environment:
+  - env
+`
+
+		var newStackYAML string
+		stdin := strings.NewReader("y")
+		var stdout bytes.Buffer
+		parent := newConfigEnvCmdForTest(stdin, &stdout, projectYAML, stackYAML, nil, nil, &newStackYAML)
+		rm := &configEnvRmCmd{parent: parent, yes: true}
+		ctx := context.Background()
+		err := rm.run(ctx, []string{"env"})
+		require.NoError(t, err)
+
+		const expectedOut = "KEY  VALUE\n"
+
+		assert.Equal(t, expectedOut, cleanStdoutIncludingPrompt(stdout.String()))
+
+		const expectedYAML = "{}\n"
+
+		assert.Equal(t, expectedYAML, newStackYAML)
+	})
+
+	t.Run("effects -> no effects", func(t *testing.T) {
+		t.Parallel()
+
+		const stackYAML = `environment:
+  - env
+  - env2
+`
 
 		env := &esc.Environment{}
 
 		var newStackYAML string
 		stdin := strings.NewReader("n")
 		var stdout bytes.Buffer
-		parent := newConfigEnvCmdForTest(stdin, &stdout, projectYAML, "", env, nil, &newStackYAML)
-		add := &configEnvAddCmd{parent: parent}
+		parent := newConfigEnvCmdForTest(stdin, &stdout, projectYAML, stackYAML, env, nil, &newStackYAML)
+		rm := &configEnvRmCmd{parent: parent}
 		ctx := context.Background()
-		err := add.run(ctx, []string{"env"})
+		err := rm.run(ctx, []string{"env2"})
 		require.Error(t, err)
 
 		const expectedOut = "KEY  VALUE\n" +
@@ -119,37 +125,9 @@ aws:region  us-west-2
 			"Save? No\n"
 
 		assert.Equal(t, expectedOut, cleanStdoutIncludingPrompt(stdout.String()))
-
-		assert.Equal(t, "", newStackYAML)
 	})
 
-	t.Run("no effects, yes", func(t *testing.T) {
-		t.Parallel()
-
-		env := &esc.Environment{}
-
-		var newStackYAML string
-		var stdout bytes.Buffer
-		parent := newConfigEnvCmdForTest(nil, &stdout, projectYAML, "", env, nil, &newStackYAML)
-		add := &configEnvAddCmd{parent: parent, yes: true}
-		ctx := context.Background()
-		err := add.run(ctx, []string{"env"})
-		require.NoError(t, err)
-
-		const expectedOut = "KEY  VALUE\n" +
-			"The stack's environment does not define the `environmentVariables`, `files`, or `pulumiConfig` properties.\n" +
-			"Without at least one of these properties, the environment will not affect the stack's behavior.\n\n"
-
-		assert.Equal(t, expectedOut, cleanStdoutIncludingPrompt(stdout.String()))
-
-		const expectedYAML = `environment:
-  - env
-`
-
-		assert.Equal(t, expectedYAML, newStackYAML)
-	})
-
-	t.Run("one import, secrets", func(t *testing.T) {
+	t.Run("two imports, secrets", func(t *testing.T) {
 		t.Parallel()
 
 		env := &esc.Environment{
@@ -162,16 +140,17 @@ aws:region  us-west-2
 		}
 
 		const stackYAML = `environment:
-  - env
+ - env
+ - env2
 `
 
 		var newStackYAML string
 		stdin := strings.NewReader("y")
 		var stdout bytes.Buffer
 		parent := newConfigEnvCmdForTest(stdin, &stdout, projectYAML, stackYAML, env, nil, &newStackYAML)
-		add := &configEnvAddCmd{parent: parent}
+		rm := &configEnvRmCmd{parent: parent}
 		ctx := context.Background()
-		err := add.run(ctx, []string{"env2"})
+		err := rm.run(ctx, []string{"env2"})
 		require.NoError(t, err)
 
 		const expectedOut = `KEY           VALUE
@@ -185,13 +164,12 @@ Save? Yes
 
 		const expectedYAML = `environment:
   - env
-  - env2
 `
 
 		assert.Equal(t, expectedYAML, newStackYAML)
 	})
 
-	t.Run("one import, secrets", func(t *testing.T) {
+	t.Run("two imports, secrets", func(t *testing.T) {
 		t.Parallel()
 
 		env := &esc.Environment{
@@ -204,16 +182,17 @@ Save? Yes
 		}
 
 		const stackYAML = `environment:
-  - env
+ - env
+ - env2
 `
 
 		var newStackYAML string
 		stdin := strings.NewReader("y")
 		var stdout bytes.Buffer
 		parent := newConfigEnvCmdForTest(stdin, &stdout, projectYAML, stackYAML, env, nil, &newStackYAML)
-		add := &configEnvAddCmd{parent: parent, showSecrets: true}
+		rm := &configEnvRmCmd{parent: parent, showSecrets: true}
 		ctx := context.Background()
-		err := add.run(ctx, []string{"env2"})
+		err := rm.run(ctx, []string{"env2"})
 		require.NoError(t, err)
 
 		const expectedOut = `KEY           VALUE
@@ -227,7 +206,6 @@ Save? Yes
 
 		const expectedYAML = `environment:
   - env
-  - env2
 `
 
 		assert.Equal(t, expectedYAML, newStackYAML)
